requirements.txt
djangorestframework==3.14.0
djangorestframework-jwt==1.0.1
docutils==0.20.1